Moved LoginDataDto to dto package.
